test: split checkpoint trigger regression module
