+++ crates/oasis7_node/src/tests_storage_replication.rs
@@ -6,6 +6,8 @@ mod storage_replication_failover_tests;
 mod storage_replication_core_tests;
 #[path = "tests_storage_replication_high_checkpoint.rs"]
 mod storage_replication_high_checkpoint_tests;
+#[path = "tests_storage_replication_first_ready_checkpoint.rs"]
+mod storage_replication_first_ready_checkpoint_tests;
 #[path = "tests_storage_replication_high_checkpoint_retained.rs"]
 mod storage_replication_high_checkpoint_retained_tests;
 #[path = "tests_storage_replication_high_checkpoint_timeout.rs"]
